discordBot/requirements.txt
discord.py
python-dotenv
asyncio
logging
sqlite3  # Optional: If using an SQLite database
